requirements.txt
discord.py
openai-whisper
